requirements.txt
dash
dash_table
plotly
numpy 
pandas
requests
dune-client
dash_bootstrap_components
datetime
requests
pytz
python-dotenv
python-binance
python-dotenv
plotly
websocket-client
dash_iconify
statsmodels
scipy
